نمایش سوالات ذخیره شده توسط مشتری و ثبت نمره توسط کارشناس قسمت ثبت نمره باید تکمیل گردد

+++ ParsKyanCrm.Application/Dtos/Users/DataFormReportDto.cs
@@ -22,8 +22,8 @@ namespace ParsKyanCrm.Application.Dtos.Users
         public int DataReportId { get; set; }
         public int RequestId { get; set; }
         public int DataFormAnswerId { get; set; }
-        public float? SystemScore { get; set; }
-        public float? AnalizeScore { get; set; }
+        public double? SystemScore { get; set; }
+        public double? AnalizeScore { get; set; }
         public string Description { get; set; }
         public byte IsActive { get; set; }
